+++ lib.php
@@ -677,10 +677,10 @@ function local_boostnavigation_get_fontawesome_icon_map() {
             $settings = explode('|', $line);
 
             // Check if there is an icon configured.
-            if (!empty($settings[6])) {
+            if (!empty($settings[7])) {
 
                 // Pick the setting which represents the icon.
-                $icon = trim($settings[6]);
+                $icon = trim($settings[7]);
 
                 // If a valid icon is given, we remember it for the iconmapping.
                 if (local_boostnavigation_verify_faicon($icon)) {